+++ src/styled/base.ts
@@ -1,11 +1,16 @@
 import { Colors } from '@blueprintjs/core'
 
+import Theme from 'Constants/theme'
 import { HighlightColor, HighlightColors } from 'Libs/Highlight'
 
 /**
  * Base theme.
  */
 export default {
+  background: {
+    [Theme.Dark]: '#293741',
+    [Theme.Light]: Colors.LIGHT_GRAY5,
+  },
   chatter: {
     avatar: {
       margin: '7px 10px 7px 0',
@@ -19,7 +24,7 @@ export default {
     Colors.GREEN5,
     Colors.ORANGE3,
     Colors.ORANGE5,
-    Colors.RED3,
+    Colors.RED4,
     Colors.RED5,
     Colors.VERMILION4,
     Colors.ROSE4,
@@ -31,7 +36,7 @@ export default {
     Colors.FOREST3,
     Colors.FOREST5,
     Colors.LIME3,
-    Colors.LIME5,
+    Colors.LIME4,
     Colors.GOLD4,
     Colors.SEPIA5,
   ],
@@ -90,13 +95,13 @@ export default {
  * Twitch username colors default mapping.
  */
 export const TwitchUserColorMap = {
-  '#0000FF': '#4966da',
-  '#008000': '#128212',
+  '#0000FF': '#6e88ef',
+  '#008000': '#23a723',
   '#00FF7F': '#0ede76',
-  '#1E90FF': '#2986e0',
-  '#8A2BE2': '#ac52ff',
-  '#9ACD32': '#8ab13b',
-  '#B22222': '#bf3434',
-  '#FF0000': '#ff2138',
-  '#FF4500': '#dc3c00',
+  '#1E90FF': '#47a4ff',
+  '#8A2BE2': '#b86cff',
+  '#9ACD32': '#97c140',
+  '#B22222': '#e64545',
+  '#FF0000': '#ff3d3d',
+  '#FF4500': '#ff6025',
 }